services/order/requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
pip==20.1.1
psycopg2==2.8.6
pytz==2021.1
setuptools==47.1.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
django-cors-headers==3.7.0
gunicorn==20.1.0